+++ src/Ext/TechnoType/Body.cpp
@@ -152,6 +152,9 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 
 	this->SelfHealGainType.Read(exINI, pSection, "SelfHealGainType");
 
+	this->InitialStrength_Cloning.Read(exINI, pSection, "InitialStrength.Cloning");
+	this->InitialStrength_Cloning_Min.Read(exINI, pSection, "InitialStrength.Cloning.Min");
+
 	// Ares 0.2
 	this->RadarJamRadius.Read(exINI, pSection, "RadarJamRadius");
 
@@ -307,6 +310,8 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Ammo_Shared)
 		.Process(this->Ammo_Shared_Group)
 		.Process(this->SelfHealGainType)
+		.Process(this->InitialStrength_Cloning)
+		.Process(this->InitialStrength_Cloning_Min)
 		;
 }
 void TechnoTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)